Move logic to AST
